Correct type of a variable to be unsigned
This caused sign extension problems leading to attempts to malloc vast
amounts of memory - under optimisation, on a 64bit system...

git-svn-id: https://svn.callweaver.org/callweaver/trunk@5608 e56c2a57-502e-0410-a6e2-86b2f1ef08c6

+++ corelib/dynstr.c
@@ -60,7 +60,7 @@ int cw_dynstr_vprintf(struct cw_dynstr **ds_p, const char *fmt, va_list ap)
 		va_list aq;
 		char *data;
 		size_t size;
-		int used;
+		size_t used;
 
 		data = NULL;
 		size = 0;
